Refactor helper methods into more semantically named classes.
