+++ src/main/java/com/nike/cerberus/auth/connector/okta/OktaAuthConnector.java
@@ -38,20 +38,24 @@ import java.util.Set;
  */
 public class OktaAuthConnector implements AuthConnector {
 
-    private final OktaAuthHelper oktaAuthHelper;
+    private final OktaApiClientHelper oktaApiClientHelper;
+
+    private final OktaClientResponseUtils oktaClientResponseUtils;
 
     @Inject
-    public OktaAuthConnector(final OktaAuthHelper oktaAuthHelper) {
+    public OktaAuthConnector(final OktaApiClientHelper oktaApiClientHelper,
+                             final OktaClientResponseUtils oktaClientResponseUtils) {
 
-        this.oktaAuthHelper = oktaAuthHelper;
+        this.oktaApiClientHelper = oktaApiClientHelper;
+        this.oktaClientResponseUtils = oktaClientResponseUtils;
     }
 
     @Override
     public AuthResponse authenticate(String username, String password) {
 
-        final AuthResult authResult = oktaAuthHelper.authenticateUser(username, password, null);
-        final String userId = oktaAuthHelper.getUserIdFromAuthResult(authResult);
-        final String userLogin = oktaAuthHelper.getUserLoginFromAuthResult(authResult);
+        final AuthResult authResult = oktaApiClientHelper.authenticateUser(username, password, null);
+        final String userId = oktaClientResponseUtils.getUserIdFromAuthResult(authResult);
+        final String userLogin = oktaClientResponseUtils.getUserLoginFromAuthResult(authResult);
 
         final AuthData authData = new AuthData()
                 .setUserId(userId)
@@ -59,18 +63,18 @@ public class OktaAuthConnector implements AuthConnector {
         final AuthResponse authResponse = new AuthResponse().setData(authData);
 
         final List<Factor> factors;
-        if (StringUtils.equals(authResult.getStatus(), OktaAuthHelper.AUTHENTICATION_MFA_REQUIRED_STATUS) ||
-                StringUtils.equals(authResult.getStatus(), OktaAuthHelper.AUTHENTICATION_MFA_ENROLL_STATUS)) {
+        if (StringUtils.equals(authResult.getStatus(), OktaClientResponseUtils.AUTHENTICATION_MFA_REQUIRED_STATUS) ||
+                StringUtils.equals(authResult.getStatus(), OktaClientResponseUtils.AUTHENTICATION_MFA_ENROLL_STATUS)) {
 
             authData.setStateToken(authResult.getStateToken());
             authResponse.setStatus(AuthStatus.MFA_REQUIRED);
 
-            factors = oktaAuthHelper.getUserFactorsFromAuthResult(authResult);
-            oktaAuthHelper.validateUserFactors(factors);
+            factors = oktaClientResponseUtils.getUserFactorsFromAuthResult(authResult);
+            oktaClientResponseUtils.validateUserFactors(factors);
 
             factors.forEach(factor -> authData.getDevices().add(new AuthMfaDevice()
                     .setId(factor.getId())
-                    .setName(oktaAuthHelper.getDeviceName(factor))));
+                    .setName(oktaClientResponseUtils.getDeviceName(factor))));
         }
         else {
             authResponse.setStatus(AuthStatus.SUCCESS);
@@ -82,9 +86,9 @@ public class OktaAuthConnector implements AuthConnector {
     @Override
     public AuthResponse mfaCheck(String stateToken, String deviceId, String otpToken) {
 
-        final AuthResult authResult = oktaAuthHelper.verifyFactor(deviceId, stateToken, otpToken);
-        final String userId = oktaAuthHelper.getUserIdFromAuthResult(authResult);
-        final String userLogin = oktaAuthHelper.getUserLoginFromAuthResult(authResult);
+        final AuthResult authResult = oktaApiClientHelper.verifyFactor(deviceId, stateToken, otpToken);
+        final String userId = oktaClientResponseUtils.getUserIdFromAuthResult(authResult);
+        final String userLogin = oktaClientResponseUtils.getUserLoginFromAuthResult(authResult);
 
         final AuthData authData = new AuthData()
                 .setUserId(userId)
@@ -101,7 +105,7 @@ public class OktaAuthConnector implements AuthConnector {
 
         Preconditions.checkNotNull(authData, "auth data cannot be null.");
 
-        final List<UserGroup> userGroups = oktaAuthHelper.getUserGroups(authData.getUserId());
+        final List<UserGroup> userGroups = oktaApiClientHelper.getUserGroups(authData.getUserId());
 
         final Set<String> groups = new HashSet<>();
         if (userGroups == null) {
