+++ src/main/java/com/nike/cerberus/auth/connector/okta/OktaClientResponseUtils.java
@@ -24,13 +24,8 @@ import com.google.inject.Inject;
 import com.nike.backstopper.apierror.ApiErrorBase;
 import com.nike.backstopper.exception.ApiException;
 import com.nike.cerberus.error.DefaultApiError;
-import com.okta.sdk.clients.AuthApiClient;
-import com.okta.sdk.clients.FactorsApiClient;
-import com.okta.sdk.clients.UserApiClient;
-import com.okta.sdk.framework.ApiClientConfiguration;
 import com.okta.sdk.models.auth.AuthResult;
 import com.okta.sdk.models.factors.Factor;
-import com.okta.sdk.models.usergroups.UserGroup;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.text.WordUtils;
 
@@ -40,9 +35,9 @@ import java.util.List;
 import java.util.Map;
 
 /**
- * Helper methods to authenticate with Okta.
+ * Created by sford4 on 1/30/17.
  */
-public class OktaAuthHelper {
+public class OktaClientResponseUtils {
 
     public static final String AUTHENTICATION_MFA_REQUIRED_STATUS = "MFA_REQUIRED";
 
@@ -58,114 +53,14 @@ public class OktaAuthHelper {
 
     private final ObjectMapper objectMapper;
 
-    private final AuthApiClient authClient;
-
-    private final UserApiClient userApiClient;
-
-    private final FactorsApiClient factorsApiClient;
-
     private final String baseUrl;
 
-    public OktaAuthHelper(final AuthApiClient authClient,
-                          final UserApiClient userApiClient,
-                          final FactorsApiClient factorsApiClient,
-                          final ObjectMapper objectMapper,
-                          final String baseUrl) {
-
-        this.authClient = authClient;
-        this.userApiClient = userApiClient;
-        this.objectMapper = objectMapper;
-        this.factorsApiClient = factorsApiClient;
-        this.baseUrl = baseUrl;
-    }
-
     @Inject
-    public OktaAuthHelper(@Named("auth.connector.okta.api_key") final String oktaApiKey,
-                          @Named("auth.connector.okta.base_url") final String baseUrl,
-                          final ObjectMapper objectMapper) {
-
-        Preconditions.checkArgument(oktaApiKey != null, "okta api key cannot be null");
-        Preconditions.checkArgument(baseUrl != null, "okta base url cannot be null");
+    public OktaClientResponseUtils(final ObjectMapper objectMapper,
+                                   @Named("auth.connector.okta.base_url") final String baseUrl) {
 
         this.objectMapper = objectMapper;
         this.baseUrl = baseUrl;
-
-        final ApiClientConfiguration clientConfiguration = new ApiClientConfiguration(baseUrl, oktaApiKey);
-        authClient = new AuthApiClient(clientConfiguration);
-        userApiClient = new UserApiClient(clientConfiguration);
-        factorsApiClient = new FactorsApiClient(clientConfiguration);
-    }
-
-    /**
-     * Request to get user group data by the user's ID.
-     *
-     * @param userId User ID
-     * @return User groups
-     */
-    protected List<UserGroup> getUserGroups(final String userId) {
-
-        try {
-            return userApiClient.getUserGroups(userId);
-        } catch (IOException ioe) {
-            final String msg = String.format("failed to get user groups for user (%s) for reason: %s", userId,
-                    ioe.getMessage());
-
-            throw ApiException.newBuilder()
-                    .withApiErrors(DefaultApiError.GENERIC_BAD_REQUEST)
-                    .withExceptionMessage(msg)
-                    .build();
-        }
-    }
-
-    /**
-     * Request for verifying a MFA factor.
-     *
-     * @param factorId MFA factor id
-     * @param stateToken State token
-     * @param passCode One Time Passcode from MFA factor
-     * @return Session login token
-     */
-    protected AuthResult verifyFactor(final String factorId,
-                                      final String stateToken,
-                                      final String passCode) {
-
-        final AuthResult authResult;
-        try {
-            authResult = authClient.authenticateWithFactor(stateToken, factorId, passCode);
-        } catch (IOException ioe) {
-            final String msg = String.format("stateToken: %s failed to verify 2nd factor for reason: %s",
-                    stateToken, ioe.getMessage());
-
-            throw ApiException.newBuilder()
-                    .withApiErrors(DefaultApiError.AUTH_BAD_CREDENTIALS)
-                    .withExceptionMessage(msg)
-                    .build();
-        }
-
-        return authResult;
-    }
-
-    /**
-     * Authenticate a user with Okta
-     *
-     * @param username   Okta username
-     * @param password   Okta password
-     * @param relayState Deep link to redirect user to after authentication
-     * @return Session login token
-     */
-    protected AuthResult authenticateUser(final String username, final String password,
-                                          final String relayState) {
-
-        try {
-            return authClient.authenticate(username, password, relayState);
-        } catch (IOException ioe) {
-            final String msg = String.format("failed to authenticate user (%s) for reason: %s", username,
-                    ioe.getMessage());
-
-            throw ApiException.newBuilder()
-                    .withApiErrors(DefaultApiError.AUTH_BAD_CREDENTIALS)
-                    .withExceptionMessage(msg)
-                    .build();        }
     }
 
     /**
@@ -183,26 +78,6 @@ public class OktaAuthHelper {
         }
 
         return WordUtils.capitalizeFully(factorProvider);
-    }
-
-    /**
-     * Get list of enrolled MFA factors for a user
-     * @param userId  Okta user ID
-     * @return List of factors
-     */
-    protected List<Factor> getFactorsByUserId(final String userId) {
-
-        Preconditions.checkArgument(userId != null, "user id cannot be null.");
-
-        try {
-            return factorsApiClient.getUserLifecycleFactors(userId);
-        } catch (IOException e) {
-            throw ApiException.newBuilder()
-                    .withApiErrors(DefaultApiError.INTERNAL_SERVER_ERROR)
-                    .withExceptionCause(e)
-                    .withExceptionMessage("Error parsing the embedded auth data from Okta.")
-                    .build();
-        }
     }
 
     /**
@@ -270,21 +145,23 @@ public class OktaAuthHelper {
      */
     protected void validateUserFactors(final List<Factor> factors) {
 
-       if(factors == null || factors.isEmpty() || factors.stream()
-               .allMatch(factor -> StringUtils.equals(factor.getStatus(), MFA_FACTOR_NOT_SETUP_STATUS)))
-       {
+        if(factors == null || factors.isEmpty() || factors.stream()
+                .allMatch(factor -> StringUtils.equals(factor.getStatus(), MFA_FACTOR_NOT_SETUP_STATUS)))
+        {
 
-           throw ApiException.newBuilder()
-                   .withApiErrors(new ApiErrorBase(
-                           DefaultApiError.MFA_SETUP_REQUIRED.getName(),
-                           DefaultApiError.MFA_SETUP_REQUIRED.getErrorCode(),
-                           "MFA is required, but user has not set up any devices in Okta.\n" +
-                                   "Please set up a MFA device in Okta: " + baseUrl,
-                           DefaultApiError.MFA_SETUP_REQUIRED.getHttpStatusCode()))
-                   .withExceptionMessage("MFA is required, but user has not set up any devices in Okta.")
-                   .build();
-       }
+            throw ApiException.newBuilder()
+                    .withApiErrors(new ApiErrorBase(
+                            DefaultApiError.MFA_SETUP_REQUIRED.getName(),
+                            DefaultApiError.MFA_SETUP_REQUIRED.getErrorCode(),
+                            "MFA is required, but user has not set up any devices in Okta.\n" +
+                                    "Please set up a MFA device in Okta: " + baseUrl,
+                            DefaultApiError.MFA_SETUP_REQUIRED.getHttpStatusCode()))
+                    .withExceptionMessage("MFA is required, but user has not set up any devices in Okta.")
+                    .build();
+        }
     }
+
+
 
     /**
      * Convenience method for parsing the Okta response and mapping it to a class.
@@ -309,5 +186,4 @@ public class OktaAuthHelper {
                     .build();
         }
     }
-
 }
