+++ src/test/java/com/nike/cerberus/auth/connector/okta/OktaMFAAuthConnectorTest.java
@@ -28,6 +28,7 @@ import com.okta.sdk.models.usergroups.UserGroupProfile;
 import org.apache.commons.lang3.StringUtils;
 import org.junit.Before;
 import org.junit.Test;
+import org.mockito.Mock;
 
 import java.util.Set;
 
@@ -37,6 +38,7 @@ import static org.mockito.Matchers.anyObject;
 import static org.mockito.Mockito.doCallRealMethod;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
+import static org.mockito.MockitoAnnotations.initMocks;
 
 /**
  * Tests the OktaMFAAuthConnector class
@@ -47,16 +49,19 @@ public class OktaMFAAuthConnectorTest {
     private OktaMFAAuthConnector oktaMFAAuthConnector;
 
     // dependencies
-    private OktaAuthHelper oktaAuthHelper;
+    @Mock
+    private OktaApiClientHelper oktaApiClientHelper;
+
+    @Mock
+    private OktaClientResponseUtils oktaClientResponseUtils;
 
     @Before
     public void setup() {
 
-        // mock dependencies
-        oktaAuthHelper = mock(OktaAuthHelper.class);
+        initMocks(this);
 
         // create test object
-        oktaMFAAuthConnector = new OktaMFAAuthConnector(oktaAuthHelper);
+        oktaMFAAuthConnector = new OktaMFAAuthConnector(oktaApiClientHelper, oktaClientResponseUtils);
     }
 
     private Factor mockFactor(String provider, String id, boolean enrolled) {
@@ -64,12 +69,12 @@ public class OktaMFAAuthConnectorTest {
         Factor factor = mock(Factor.class);
         when(factor.getId()).thenReturn(id);
         when(factor.getProvider()).thenReturn(provider);
-        when(oktaAuthHelper.getDeviceName(factor)).thenCallRealMethod();
+        when(oktaClientResponseUtils.getDeviceName(factor)).thenCallRealMethod();
 
         if (enrolled) {
             when(factor.getStatus()).thenReturn("status");
         } else {
-            when(factor.getStatus()).thenReturn(OktaAuthHelper.MFA_FACTOR_NOT_SETUP_STATUS);
+            when(factor.getStatus()).thenReturn(OktaClientResponseUtils.MFA_FACTOR_NOT_SETUP_STATUS);
         }
 
         return factor;
@@ -89,12 +94,12 @@ public class OktaMFAAuthConnectorTest {
         String id = "id";
 
         AuthResult authResult = mock(AuthResult.class);
-        when(authResult.getStatus()).thenReturn(OktaAuthHelper.AUTHENTICATION_SUCCESS_STATUS);
+        when(authResult.getStatus()).thenReturn(OktaClientResponseUtils.AUTHENTICATION_SUCCESS_STATUS);
 
-        when(oktaAuthHelper.authenticateUser(username, password, null)).thenReturn(authResult);
-        when(oktaAuthHelper.getUserIdFromAuthResult(authResult)).thenReturn(id);
-        when(oktaAuthHelper.getUserLoginFromAuthResult(authResult)).thenReturn(email);
-        doCallRealMethod().when(oktaAuthHelper).validateUserFactors(anyObject());
+        when(oktaApiClientHelper.authenticateUser(username, password, null)).thenReturn(authResult);
+        when(oktaClientResponseUtils.getUserIdFromAuthResult(authResult)).thenReturn(id);
+        when(oktaClientResponseUtils.getUserLoginFromAuthResult(authResult)).thenReturn(email);
+        doCallRealMethod().when(oktaClientResponseUtils).validateUserFactors(anyObject());
 
         // do the call
         AuthResponse result = this.oktaMFAAuthConnector.authenticate(username, password);
@@ -119,11 +124,11 @@ public class OktaMFAAuthConnectorTest {
         AuthResult authResult = mock(AuthResult.class);
         when(authResult.getStateToken()).thenReturn("state token");
 
-        when(oktaAuthHelper.authenticateUser(username, password, null)).thenReturn(authResult);
-        when(authResult.getStatus()).thenReturn(OktaAuthHelper.AUTHENTICATION_MFA_REQUIRED_STATUS);
-        when(oktaAuthHelper.getUserIdFromAuthResult(authResult)).thenReturn(id);
-        when(oktaAuthHelper.getUserLoginFromAuthResult(authResult)).thenReturn(email);
-        when(oktaAuthHelper.getUserFactorsFromAuthResult(authResult)).thenReturn(Lists.newArrayList(factor));
+        when(oktaApiClientHelper.authenticateUser(username, password, null)).thenReturn(authResult);
+        when(authResult.getStatus()).thenReturn(OktaClientResponseUtils.AUTHENTICATION_MFA_REQUIRED_STATUS);
+        when(oktaClientResponseUtils.getUserIdFromAuthResult(authResult)).thenReturn(id);
+        when(oktaClientResponseUtils.getUserLoginFromAuthResult(authResult)).thenReturn(email);
+        when(oktaClientResponseUtils.getUserFactorsFromAuthResult(authResult)).thenReturn(Lists.newArrayList(factor));
 
         // do the call
         AuthResponse result = this.oktaMFAAuthConnector.authenticate(username, password);
@@ -148,12 +153,12 @@ public class OktaMFAAuthConnectorTest {
         AuthResult authResult = mock(AuthResult.class);
         when(authResult.getStateToken()).thenReturn("state token");
 
-        when(oktaAuthHelper.authenticateUser(username, password, null)).thenReturn(authResult);
-        when(authResult.getStatus()).thenReturn(OktaAuthHelper.AUTHENTICATION_MFA_ENROLL_STATUS);
-        when(oktaAuthHelper.getUserIdFromAuthResult(authResult)).thenReturn(id);
-        when(oktaAuthHelper.getUserLoginFromAuthResult(authResult)).thenReturn(email);
-        when(oktaAuthHelper.getFactorsByUserId(id)).thenReturn(Lists.newArrayList());
-        doCallRealMethod().when(oktaAuthHelper).validateUserFactors(anyObject());
+        when(oktaApiClientHelper.authenticateUser(username, password, null)).thenReturn(authResult);
+        when(authResult.getStatus()).thenReturn(OktaClientResponseUtils.AUTHENTICATION_MFA_ENROLL_STATUS);
+        when(oktaClientResponseUtils.getUserIdFromAuthResult(authResult)).thenReturn(id);
+        when(oktaClientResponseUtils.getUserLoginFromAuthResult(authResult)).thenReturn(email);
+        when(oktaApiClientHelper.getFactorsByUserId(id)).thenReturn(Lists.newArrayList());
+        doCallRealMethod().when(oktaClientResponseUtils).validateUserFactors(anyObject());
 
         // do the call
         AuthResponse result = this.oktaMFAAuthConnector.authenticate(username, password);
@@ -179,11 +184,11 @@ public class OktaMFAAuthConnectorTest {
         AuthResult authResult = mock(AuthResult.class);
         when(authResult.getStateToken()).thenReturn("state token");
 
-        when(oktaAuthHelper.authenticateUser(username, password, null)).thenReturn(authResult);
-        when(authResult.getStatus()).thenReturn(OktaAuthHelper.AUTHENTICATION_SUCCESS_STATUS);
-        when(oktaAuthHelper.getUserIdFromAuthResult(authResult)).thenReturn(id);
-        when(oktaAuthHelper.getUserLoginFromAuthResult(authResult)).thenReturn(email);
-        when(oktaAuthHelper.getFactorsByUserId(id)).thenReturn(Lists.newArrayList(factor));
+        when(oktaApiClientHelper.authenticateUser(username, password, null)).thenReturn(authResult);
+        when(authResult.getStatus()).thenReturn(OktaClientResponseUtils.AUTHENTICATION_SUCCESS_STATUS);
+        when(oktaClientResponseUtils.getUserIdFromAuthResult(authResult)).thenReturn(id);
+        when(oktaClientResponseUtils.getUserLoginFromAuthResult(authResult)).thenReturn(email);
+        when(oktaApiClientHelper.getFactorsByUserId(id)).thenReturn(Lists.newArrayList(factor));
 
         // do the call
         AuthResponse result = this.oktaMFAAuthConnector.authenticate(username, password);
@@ -207,9 +212,9 @@ public class OktaMFAAuthConnectorTest {
         String id = "id";
 
         AuthResult authResult = mock(AuthResult.class);
-        when(oktaAuthHelper.verifyFactor(deviceId, stateToken, otpToken)).thenReturn(authResult);
-        when(oktaAuthHelper.getUserIdFromAuthResult(authResult)).thenReturn(id);
-        when(oktaAuthHelper.getUserLoginFromAuthResult(authResult)).thenReturn(email);
+        when(oktaApiClientHelper.verifyFactor(deviceId, stateToken, otpToken)).thenReturn(authResult);
+        when(oktaClientResponseUtils.getUserIdFromAuthResult(authResult)).thenReturn(id);
+        when(oktaClientResponseUtils.getUserLoginFromAuthResult(authResult)).thenReturn(email);
 
         // do the call
         AuthResponse result = this.oktaMFAAuthConnector.mfaCheck(stateToken, deviceId, otpToken);
@@ -238,7 +243,7 @@ public class OktaMFAAuthConnectorTest {
         when(profile2.getName()).thenReturn(name2);
         when(group2.getProfile()).thenReturn(profile2);
 
-        when(oktaAuthHelper.getUserGroups(id)).thenReturn(Lists.newArrayList(group1, group2));
+        when(oktaApiClientHelper.getUserGroups(id)).thenReturn(Lists.newArrayList(group1, group2));
 
         // do the call
         Set<String> result = this.oktaMFAAuthConnector.getGroups(authData);
